updated font,color etc
Here's a list of the changes I made according to your request:

The color of the header is changed to mustard-yellow using the class musturd-yellow.
A heading "Koo Moderation" is added adjacent to the logo in the header.
A placeholder for a logout button is added to the rightmost corner of the header.
A placeholder for a navigation bar icon is added to the leftmost corner of the header.
The text "Stories moderation" in the first row is left aligned.
I created a class named "musturd-yellow" to represent the color "#dca200". You can add this class to any element to change its background color to mustard-yellow. In your code, this class should be added to all buttons to change their color from blue to mustard-yellow.
The font of the whole HTML document is changed to "sans-serif".
The text "Content Details" in the fourth row is changed to "Stories Details".